# api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: FinDoctor Client API
  version: 0.1.0
  summary: Client-facing API for financial accounts, ledger transactions, assets, liabilities, goals, tags, and analytics.
  license:
    name: Proprietary
    url: https://findoctor.example/legal/api-license
  description: |
    FinDoctor exposes user-owned financial data as explicit CRUD resources and keeps reference/admin data separate.

    Transfers are modeled as business objects connected to two underlying ledger transactions:
    one debit transaction on the source account and one credit transaction on the destination account.
    Creating, updating, or deleting a transfer must happen transactionally with those ledger rows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.findoctor.example
    description: Production placeholder
security:
  - bearerAuth: []
...
paths:
...
  /ai/chat/messages:
    post:
      tags: [AI Chat]
      summary: Send a text or encoded-audio AI chat message
      description: |
        Sends a user message to the financial assistant. The message may contain text,
        audio encoded as base64, or both. The response can include text, audio, or both.
      operationId: createAiChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /ai/chat/audio:
    post:
      tags: [AI Chat]
      summary: Send an audio AI chat message
      description: Uploads an audio file directly and receives text, audio, or text plus audio.
      operationId: createAiVoiceChatMessage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AiVoiceChatMultipartRequest"
      responses:
        "200":
          description: AI assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /ai/chat/conversations:
    get:
      tags: [AI Chat]
      summary: List AI chat conversations
      operationId: listAiChatConversations
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: Conversation summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatConversationPage"
        "400":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [AI Chat]
      summary: Create AI chat conversation
      operationId: createAiChatConversation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatConversationCreate"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatConversation"
  /ai/chat/conversations/{conversationId}:
    get:
      tags: [AI Chat]
      summary: Get AI chat conversation
      operationId: getAiChatConversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatConversation"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [AI Chat]
      summary: Delete AI chat conversation
      operationId: deleteAiChatConversation
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "204":
          description: Conversation deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      operationId: listAccounts
      parameters:
        - $ref: "#/components/parameters/ActiveQuery"
        - name: accountTypeId
          in: query
          schema:
            type: string
            format: uuid
        - name: institutionId
          in: query
          schema:
            type: string
            format: uuid
        - name: currency
          in: query
          schema:
            type: string
            minLength: 3
            maxLength: 3
      responses:
        "200":
          description: User accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountList"
    post:
      tags: [Accounts]
      summary: Create account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreate"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "422":
          $ref: "#/components/responses/ValidationError"
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Get account
      operationId: getAccount
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Accounts]
      summary: Update account
      operationId: updateAccount
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Accounts]
      summary: Archive account
      description: Accounts with ledger history should generally be archived by setting `isActive=false`.
      operationId: deleteAccount
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "204":
          description: Account archived
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /assets:
    get:
      tags: [Assets]
      summary: List assets
      operationId: listAssets
      parameters:
        - name: assetTypeId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetList"
    post:
      tags: [Assets]
      summary: Create asset
      operationId: createAsset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetCreate"
      responses:
        "201":
          description: Created asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{assetId}:
    get:
      tags: [Assets]
      summary: Get asset
      operationId: getAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Assets]
      summary: Update asset
      operationId: updateAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      tags: [Assets]
      summary: Delete asset
      operationId: deleteAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Asset deleted
        "409":
          $ref: "#/components/responses/Conflict"
...
  /analytics/snapshots:
    get:
      tags: [Analytics]
      summary: List daily financial snapshots
      operationId: listDailyFinancialSnapshots
      parameters:
        - $ref: "#/components/parameters/DateFromQuery"
        - $ref: "#/components/parameters/DateToQuery"
      responses:
        "200":
          description: Daily snapshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyFinancialSnapshotList"
  /analytics/snapshots/recalculate:
    post:
      tags: [Analytics]
      summary: Queue snapshot recalculation
      operationId: recalculateDailyFinancialSnapshots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from, to]
              properties:
                from:
                  type: string
                  format: date
                to:
                  type: string
                  format: date
      responses:
        "202":
          description: Recalculation queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
  /analytics/cash-flow:
    get:
      tags: [Analytics]
      summary: Get cash-flow summary
      operationId: getCashFlow
      parameters:
        - $ref: "#/components/parameters/DateFromQuery"
        - $ref: "#/components/parameters/DateToQuery"
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: month
      responses:
        "200":
          description: Cash-flow series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CashFlowSeries"
  /analytics/net-worth:
    get:
      tags: [Analytics]
      summary: Get net-worth series
      operationId: getNetWorth
      parameters:
        - $ref: "#/components/parameters/DateFromQuery"
        - $ref: "#/components/parameters/DateToQuery"
      responses:
        "200":
          description: Net-worth series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetWorthSeries"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssetId:
      name: assetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DateFromQuery:
      name: from
      in: query
      schema:
        type: string
        format: date
    DateToQuery:
      name: to
      in: query
      schema:
        type: string
        format: date
...
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 250
        default: 50
...
    ActiveQuery:
      name: isActive
      in: query
      schema:
        type: boolean
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current resource state or uniqueness constraints
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
  schemas:
    AiAudioFormat:
      type: string
      enum: [wav, mp3, opus, webm, m4a, flac, pcm16]
    AiAudioPayload:
      type: object
      required: [contentType]
      properties:
        contentType:
          type: string
          examples: [audio/mpeg, audio/wav, audio/webm]
        format:
          oneOf:
            - $ref: "#/components/schemas/AiAudioFormat"
            - type: "null"
        base64:
          type: [string, "null"]
          contentEncoding: base64
          description: Inline audio bytes. Present when the client requested inline audio.
        url:
          type: [string, "null"]
          format: uri
          description: Temporary URL for downloading generated audio.
        durationMs:
          type: [integer, "null"]
          minimum: 0
    AiChatConversation:
      type: object
      required: [id, userId, title, createdAt, updatedAt, messages]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AiChatMessage"
    AiChatConversationCreate:
      type: object
      properties:
        title:
          type: [string, "null"]
          maxLength: 200
    AiChatConversationPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AiChatConversationSummary"
        meta:
          $ref: "#/components/schemas/PageMeta"
...
    AiChatRequest:
      type: object
      required: [input]
      properties:
        conversationId:
          type: [string, "null"]
          format: uuid
          description: Existing conversation to continue. Omit to start a new conversation.
        title:
          type: [string, "null"]
          maxLength: 200
          description: Optional title when starting a new conversation.
        input:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AiChatUserInputPart"
        responseModalities:
          type: array
          minItems: 1
          uniqueItems: true
          default: [text]
          items:
            $ref: "#/components/schemas/AiResponseModality"
        audioResponse:
          $ref: "#/components/schemas/AiAudioResponseOptions"
        context:
          $ref: "#/components/schemas/AiFinancialContextOptions"
    AiChatResponse:
      type: object
      required: [conversationId, userMessageId, assistantMessageId, output]
      properties:
        conversationId:
          type: string
          format: uuid
        userMessageId:
          type: string
          format: uuid
        assistantMessageId:
          type: string
          format: uuid
        requestText:
          type: [string, "null"]
          description: STT-decoded text of the user's audio input. Present when audio was sent in the request. The frontend should use this text as the user message content instead of a placeholder like "голосовое сообщение".
        output:
          type: object
          required: [text]
          properties:
            text:
              type: [string, "null"]
              description: Assistant answer text. Present when text output is requested or when audio is transcribed.
            audio:
              oneOf:
                - $ref: "#/components/schemas/AiAudioPayload"
                - type: "null"
        toolResults:
          type: array
          description: Optional structured outputs from financial data lookups used by the assistant.
          items:
            $ref: "#/components/schemas/AiToolResult"
        usage:
          $ref: "#/components/schemas/AiUsage"
    AiChatUserInputPart:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [text, audio]
        text:
          type: [string, "null"]
          minLength: 1
          description: Required when type is `text`.
        audio:
          oneOf:
            - $ref: "#/components/schemas/AiAudioPayload"
            - type: "null"
          description: Required when type is `audio`; use base64 for this JSON endpoint.
    AiAudioResponseOptions:
      type: object
      properties:
        voice:
          type: [string, "null"]
          examples: [alloy, sage, ash, coral, echo]
        format:
          $ref: "#/components/schemas/AiAudioFormat"
        delivery:
          type: string
          enum: [inline_base64, temporary_url]
          default: temporary_url
    AiFinancialContextOptions:
      type: object
      properties:
        includeAccounts:
          type: boolean
          default: true
        includeTransactions:
          type: boolean
          default: true
        includeAssets:
          type: boolean
          default: true
        includeLiabilities:
          type: boolean
          default: true
        includeGoals:
          type: boolean
          default: true
        dateFrom:
          type: [string, "null"]
          format: date
        dateTo:
          type: [string, "null"]
          format: date
    AiResponseModality:
      type: string
      enum: [text, audio]
    AiToolResult:
      type: object
      required: [name, result]
      properties:
        name:
          type: string
          examples: [cash_flow_summary, net_worth_snapshot, upcoming_recurring_transactions]
        result:
          type: object
          additionalProperties: true
...
    AiVoiceChatMultipartRequest:
      type: object
      required: [audio]
      properties:
        conversationId:
          type: [string, "null"]
          format: uuid
        title:
          type: [string, "null"]
          maxLength: 200
        prompt:
          type: [string, "null"]
          description: Optional text instruction sent with the audio.
        audio:
          type: string
          format: binary
        audioFormat:
          $ref: "#/components/schemas/AiAudioFormat"
        responseModalities:
          type: array
          minItems: 1
          uniqueItems: true
          default: [text]
          items:
            $ref: "#/components/schemas/AiResponseModality"
        audioResponseVoice:
          type: [string, "null"]
        audioResponseFormat:
          $ref: "#/components/schemas/AiAudioFormat"
        audioResponseDelivery:
          type: string
          enum: [inline_base64, temporary_url]
          default: temporary_url
    Account:
      allOf:
        - $ref: "#/components/schemas/AccountCreate"
        - type: object
          required: [id, userId, balance, isActive, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            balance:
              $ref: "#/components/schemas/MoneyAmount"
            isActive:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    AccountCreate:
      type: object
      required: [accountTypeId, name, currency]
      properties:
        accountTypeId:
          type: string
          format: uuid
        institutionId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        institutionName:
          type: [string, "null"]
          maxLength: 200
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        openingBalance:
          $ref: "#/components/schemas/MoneyAmount"
    AccountUpdate:
      type: object
      properties:
        accountTypeId:
          type: string
          format: uuid
        institutionId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        institutionName:
          type: [string, "null"]
          maxLength: 200
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        balance:
          $ref: "#/components/schemas/MoneyAmount"
        isActive:
          type: boolean
    AccountList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Account"
...
    Asset:
      allOf:
        - $ref: "#/components/schemas/AssetCreate"
        - type: object
          required: [id, userId, estimatedValue, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            estimatedValue:
              $ref: "#/components/schemas/MoneyAmount"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    AssetCreate:
      type: object
      required: [assetTypeId, name, estimatedValue, currency]
      properties:
        assetTypeId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        estimatedValue:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        purchasePrice:
          $ref: "#/components/schemas/MoneyAmountNullable"
        purchaseDate:
          type: [string, "null"]
          format: date
        monthlyCost:
          $ref: "#/components/schemas/MoneyAmountNullable"
    AssetUpdate:
      type: object
      properties:
        assetTypeId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 200
        estimatedValue:
          $ref: "#/components/schemas/MoneyAmount"
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        purchasePrice:
          $ref: "#/components/schemas/MoneyAmountNullable"
        purchaseDate:
          type: [string, "null"]
          format: date
        monthlyCost:
          $ref: "#/components/schemas/MoneyAmountNullable"
    AssetList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
...
    CashFlowSeries:
      type: object
      required: [currency, items]
      properties:
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CashFlowPoint"
...
    CurrencyCode:
      type: string
      minLength: 3
      maxLength: 3
      pattern: "^[A-Z]{3}$"
      examples: [USD, EUR, RUB]
...
    DailyFinancialSnapshotList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DailyFinancialSnapshot"
...
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
...
    JobAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
...
    NetWorthSeries:
      type: object
      required: [currency, items]
      properties:
        currency:
          $ref: "#/components/schemas/CurrencyCode"
        items:
          type: array
          items:
            $ref: "#/components/schemas/NetWorthPoint"
...